requirements.txt
matplotlib==3.5.3
numpy==1.22.4
opencv_python_headless==4.6.0.66
pandas==1.4.3
Pillow==9.5.0
segmentation_models_pytorch==0.3.3
streamlit==1.23.1
torch==2.0.1
torchvision==0.15.2
